Replaced array_merge with array_merge_recursive in register_post_type()
This fixes an issue where values set in the 2nd dimension array 'labels'
were not being merged because array_merge only merges the 1st dimension
array elements.

The fix allows the user to customize a single element in the 'labels'
array and ensure that the Super CPT defaults apply to the remaining
sibling elements rather than have the WordPress defaults take precendence.